ethos-pulse/pipeline/requirements.txt
pandas==2.2.2
openpyxl==3.1.4
python-dotenv==1.0.1
numpy==1.26.4
